# Reduce string allocations when writing documents