--- a/OpenXmlFormats/Wordprocessing/HdrFtr.cs
+++ b/OpenXmlFormats/Wordprocessing/HdrFtr.cs
@@ -262,7 +262,7 @@ namespace NPOI.OpenXmlFormats.Wordprocessing
                 else if (o is CT_Markup)
                     ((CT_Markup)o).Write(sw, "customXmlInsRangeEnd");
             }
-            sw.Write(string.Format("</w:{0}>", nodeName));
+            sw.WriteEndW(nodeName);
         }
 
         public ArrayList Items
@@ -648,7 +648,7 @@ namespace NPOI.OpenXmlFormats.Wordprocessing
                 this.numStart.Write(sw, "numStart");
             if (this.numRestart != null)
                 this.numRestart.Write(sw, "numRestart");
-            sw.Write(string.Format("</w:{0}>", nodeName));
+            sw.WriteEndW(nodeName);
         }
 
         [XmlElement(Order = 0)]
@@ -728,7 +728,7 @@ namespace NPOI.OpenXmlFormats.Wordprocessing
             sw.Write(string.Format("<w:{0}", nodeName));
             XmlHelper.WriteAttribute(sw, "w:val", this.val.ToString());
             sw.Write(">");
-            sw.Write(string.Format("</w:{0}>", nodeName));
+            sw.WriteEndW(nodeName);
         }
 
         private ST_FtnPos valField;
@@ -1098,7 +1098,7 @@ namespace NPOI.OpenXmlFormats.Wordprocessing
                     ((CT_CustomXmlBlock)o).Write(sw, "customXml");
                 i++;
             }
-            sw.Write(string.Format("</w:{0}>", nodeName));
+            sw.WriteEndW(nodeName);
         }
 
 
@@ -1523,7 +1523,7 @@ namespace NPOI.OpenXmlFormats.Wordprocessing
                     x.Write(sw, "footnote");
                 }
             }
-            sw.Write(string.Format("</w:{0}>", nodeName));
+            sw.WriteEndW(nodeName);
         }
 
         [XmlElement("footnote", Order = 0)]
@@ -1635,7 +1635,7 @@ namespace NPOI.OpenXmlFormats.Wordprocessing
                 this.numStart.Write(sw, "numStart");
             if (this.numRestart != null)
                 this.numRestart.Write(sw, "numRestart");
-            sw.Write(string.Format("</w:{0}>", nodeName));
+            sw.WriteEndW(nodeName);
         }
 
         [XmlElement(Order = 0)]
@@ -1715,7 +1715,7 @@ namespace NPOI.OpenXmlFormats.Wordprocessing
             sw.Write(string.Format("<w:{0}", nodeName));
             XmlHelper.WriteAttribute(sw, "w:val", this.val.ToString());
             sw.Write(">");
-            sw.Write(string.Format("</w:{0}>", nodeName));
+            sw.WriteEndW(nodeName);
         }
 
         private ST_EdnPos valField;
@@ -1797,7 +1797,7 @@ namespace NPOI.OpenXmlFormats.Wordprocessing
                     x.Write(sw, "endnote");
                 }
             }
-            sw.Write(string.Format("</w:{0}>", nodeName));
+            sw.WriteEndW(nodeName);
         }
 
         private List<CT_FtnEdnSepRef> endnoteField;
@@ -1857,7 +1857,7 @@ namespace NPOI.OpenXmlFormats.Wordprocessing
             XmlHelper.WriteAttribute(sw, "w:customMarkFollows", this.customMarkFollows.ToString());
             XmlHelper.WriteAttribute(sw, "w:id", this.id);
             sw.Write(">");
-            sw.Write(string.Format("</w:{0}>", nodeName));
+            sw.WriteEndW(nodeName);
         }
 
         [XmlAttribute(Form = System.Xml.Schema.XmlSchemaForm.Qualified)]