`string.Format` is a bad thing in hot paths. Changed code to call multiple `Write`s instead of string concatenation.

## NPOI.Benchmarks.LargeExcelFileBenchmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |Load|11.741 s|1.1257 s|3.65 GB|
| **New** |	| **11.626 s (-1%)** | **2.0913 s** | **3732.62 MB (0%)** |
| Old |Write|3.483 s|0.1738 s|1.19 GB|
| **New** |	| **3.058 s (-12%)** | **0.2296 s** | **611.83 MB (-50%)** |
| Old |Evaluate|62.677 s|4.1891 s|75.32 GB|
| **New** |	| **64.658 s (+3%)** | **24.1520 s** | **77128.7 MB (0%)** |


## NPOI.Benchmarks.RangeValuesBenchmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |Double|85.46 ms|0.367 ms|31.45 MB|
| **New** |	| **82.83 ms (-3%)** | **0.382 ms** | **26.98 MB (-14%)** |
| Old |String|99.93 ms|0.449 ms|59.32 MB|
| **New** |	| **91.75 ms (-8%)** | **0.803 ms** | **53.92 MB (-9%)** |
| Old |Date|114.71 ms|1.603 ms|35.73 MB|
| **New** |	| **108.22 ms (-6%)** | **0.239 ms** | **30.55 MB (-14%)** |
| Old |Formulas|1,064.12 ms|10.889 ms|1480.15 MB|
| **New** |	| **1,057.00 ms (-1%)** | **4.548 ms** | **1473.42 MB (0%)** |