@tonyqus 2023-09-06
I'm checking if we can rewrite serialization part with https://github.com/Cysharp/ZString. 

According to [nuget stats](https://www.nuget.org/packages/ZString), this library looks popular. 

I found this library via [this post](https://neuecc.medium.com/modern-high-performance-c-2023-edition-cd9a81838ad8)