@lahma 2023-09-06
I can recommend that library, Orchard Core for example uses that, got a PR approved once to that lib https://github.com/Cysharp/ZString/pull/50 .